requirements.txt
streamlit
pandas
ibm-watson-machine-learning
ibm-cos-sdk
fpdf2
matplotlib
